requirements.txt
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.0
python-dateutil>=2.8.1
requests-cache>=0.6.0
tqdm>=4.61.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
Flask>=2.0.0
Flask-CORS>=3.0.0
